+++ docs/changelog.md
@@ -11,6 +11,12 @@ See the [GitHub "Features" project](https://github.com/users/peter88213/projects
 
 ---
 
+### Version 0.4.0
+
+- Renamed the "Substitutions" submenu to "Options" and placed it in the new "Tools" menu.
+- New option: "Large toolbar icons".
+
+
 ### Version 0.3.0
 
 - The configuration is persistent now. Closes #2
@@ -28,6 +34,6 @@ See the [GitHub "Features" project](https://github.com/users/peter88213/projects
 
 ### Version 0.1.0
 
-Alpha version based on the nvtlview library version 5.2.4.
+Alpha version based on the *nvtlview* library version 5.2.4.
 
 To do: Before changing or closing the project, ask for saving changes.